+++ src/screens/Main/components/Home/CPU.tsx
@@ -6,9 +6,17 @@ import LogoAmd from "./../../../../assets/cpu_vendor_logo/amd.png";
 import LogoArm from "./../../../../assets/cpu_vendor_logo/arm.png";
 import { useTranslation } from "react-i18next";
 import { Image, View } from "react-native";
+import moment from "moment";
+import { useMemo } from "react";
 
 export default function HomeCPU({ d }: { d: ServerStats }) {
     const { t } = useTranslation();
+
+    const upTime = useMemo(() => {
+        const dur = moment.duration(d.uptime, 'seconds');
+
+        return `${dur.days()} ${t('days')} ${dur.hours()} ${t('hours')} ${dur.minutes()} ${t('minutes')} ${dur.seconds()} ${t('seconds')}`
+    }, []);
     return (
         <>
             {d.cpu.name.startsWith("Intel") ? <LogoIntel width={150} height={80} />
@@ -22,17 +30,17 @@ export default function HomeCPU({ d }: { d: ServerStats }) {
                 <View style={{ display: 'flex', flexDirection: 'row' }}>
                     <View style={{ flex: 1 }}>
                         <Text variant="labelLarge" style={{ color: 'gray' }}>{t('cpu_clock')}</Text>
-                        <Text>{d.cpu.speed} MHz</Text>
+                        <Text>{d.cpu.speed.length > 1 ? d.cpu.speed.length : "(unknown) 0"} MHz</Text>
                     </View>
                 </View>
                 <View style={{ display: 'flex', flexDirection: 'row' }}>
                     <View style={{ flex: 1 }}>
                         <Text variant="labelLarge" style={{ color: 'gray' }}>{t('architecture')}</Text>
-                        <Text>{d.cpu.architecture}</Text>
+                        <Text>{d.cpu.architecture ?? "unknown"}</Text>
                     </View>
                     <View style={{ flex: 1 }}>
                         <Text variant="labelLarge" style={{ color: 'gray' }}>{t('cache_size')}</Text>
-                        <Text>{d.cpu.cache_size}</Text>
+                        <Text>{d.cpu.cache_size.length > 0 ? d.cpu.cache_size : "unknown"}</Text>
                     </View>
                 </View>
                 <View style={{ display: 'flex', flexDirection: 'row' }}>
@@ -45,8 +53,14 @@ export default function HomeCPU({ d }: { d: ServerStats }) {
                         <Text>{d.cpu.physical_cores}</Text>
                     </View>
                 </View>
-                <Text variant="labelLarge" style={{ color: 'gray' }}>{t('flags')}</Text>
-                <Text>{d.cpu.flags}</Text>
+                <View style={{ flex: 1 }}>
+                    <Text variant="labelLarge" style={{ color: 'gray' }}>{t('flags')}</Text>
+                    <Text>{d.cpu.flags.length > 0 ? d.cpu.flags : "unknown"}</Text>
+                </View>
+                <View style={{ flex: 1 }}>
+                    <Text variant="labelLarge" style={{ color: 'gray' }}>{t('system_uptime')}</Text>
+                    <Text style={{ textTransform: "lowercase" }}>{upTime}</Text>
+                </View>
             </View>
         </>
     )